fix mail message initialization (#46)
When multiple mails are being sent (e.g. admin, user), TYPO3Mailer
currently reuses the MailMessage instance. In effect, fields are not
reset and headers such as CC are reused in additional mails.

This patch creates a new MailMessage instance after init is called. The
additional initialization of emailObj in __construct is not removed to
leave the constructor with the field initialized.